@@ -3,7 +3,7 @@ build-backend = "flit_core.buildapi"
 requires = ["flit_core >=3.2,<4"]
 
 [project]
-authors = [{name = "Thomas Krapp", email = "t.krapp@gmail.com"}]
+authors = [{ name = "Thomas Krapp", email = "t.krapp@gmail.com" }]
 classifiers = [
   "License :: OSI Approved :: MIT License",
   "Environment :: Web Environment",
@@ -23,16 +23,16 @@ classifiers = [
 ]
 dependencies = ["Django >= 3.0"]
 dynamic = ["version", "description"]
-license = {file = "LICENSE"}
+license = { file = "LICENSE" }
 name = "typed_app_settings"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 
 [project.urls]
 Home = "https://github.com/tkrapp/django-typed-app-settings"
 
 [project.optional-dependencies]
-dev = ["black", "pytest"]
+dev = ["black"]
 
 [tool.isort]
 profile = "black"
