@@ -1,15 +1,9 @@
 import os
 import pathlib
-import sys
 import types
 import unittest
 from collections import deque
-
-if sys.version_info <= (3, 8):
-    from typing import Sequence
-else:
-    from collections.abc import Sequence
-
+from collections.abc import Sequence
 from typing import Any, Type, Union
 
 from django.conf import settings as django_settings
